requirements.txt
fastapi==0.110.2
uvicorn==0.29.0
huggingface-hub==0.22.2
ctransformers==0.2.27
langchain==0.1.14
langchain-community==0.0.28
langchain-core==0.1.50
langchain-huggingface==0.0.5
langchain-pinecone==0.0.6
langchain-text-splitters==0.0.2
pinecone==2.2.4
pypdf==3.17.4
sentence-transformers==2.2.2
python-dotenv==1.0.1
requests==2.31.0
numpy==1.26.4
tqdm==4.66.1
jinja2==3.1.2
starlette==0.36.3
python-multipart==0.0.7
safetensors==0.4.0
tokenizers==0.14.1
